feat!: Remove deprecated execution message (#54)
This removes the execution message and adds a comment to the README on
how to readd them via Autocommands.

Also I implemented a function to handle breaking changes. It can later
be used for https://github.com/okuuva/auto-save.nvim/pull/48

This merges in a feature branch. I would like to collect all potential
breaking changes there until we can release it all together in a version
1.0.0

---------

Co-authored-by: okuuva <okuuva@users.noreply.github.com>